backend/pyproject.toml
[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[project]
name = "hermes-desktop-daemon"
version = "0.1.0"
description = "Headless daemon for Hermes Desktop v1"
requires-python = ">=3.11"
dependencies = [
  "aiohttp>=3.9,<4",
  "pydantic>=2.7,<3",
  "python-dotenv>=1.0,<2",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
